Put laws in combined types in their own trait.
