+++ tests/src/test/scala/spire/laws/LawTests.scala
@@ -36,18 +36,18 @@ class LawTests extends FunSuite with Discipline {
   checkAll("Int",        RingLaws[Int].cRing)
   checkAll("Long",       RingLaws[Long].cRing)
   checkAll("BigInt",     RingLaws[BigInt].euclideanRing)
-  checkAll("BigInt",     BaseLaws[BigInt].signedGCDRing)
+  checkAll("BigInt",     CombinationLaws[BigInt].signedGCDRing)
   checkAll("BigInteger", RingLaws[BigInteger].euclideanRing)
-  checkAll("BigInteger",     BaseLaws[BigInteger].signedGCDRing)
+  checkAll("BigInteger", CombinationLaws[BigInteger].signedGCDRing)
   checkAll("Rational",   RingLaws[Rational].field)
-  checkAll("Rational",   BaseLaws[BigInt].signedGCDRing)
+  checkAll("Rational",   CombinationLaws[BigInt].signedGCDRing)
   checkAll("Real",       RingLaws[Real].field)
   checkAll("UByte",      RingLaws[UByte].cRig)
   checkAll("UShort",     RingLaws[UShort].cRig)
   checkAll("UInt",       RingLaws[UInt].cRig)
   checkAll("ULong",      RingLaws[ULong].cRig)
   checkAll("SafeLong",   RingLaws[SafeLong].euclideanRing)
-  checkAll("SafeLong",   BaseLaws[SafeLong].signedGCDRing)
+  checkAll("SafeLong",   CombinationLaws[SafeLong].signedGCDRing)
 
   checkAll("Complex[Rational]", RingLaws[Complex[Rational]].field)
